src/web/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
psycopg[binary]>=3.2.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
PyYAML>=6.0.1
requests>=2.31.0
json-repair>=0.28.0
PyMySQL>=1.1.1
pytest>=8.2.0
